Desktop/careerhq-agents/requirements.txt
# Core dependencies
langchain>=0.1.8
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.24

# LLM providers
openai>=1.69.0

# Web framework
fastapi>=0.109.2
uvicorn>=0.27.1
sse-starlette>=1.6.5

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.1
json-repair>=0.6.0

# Data handling
pandas>=2.2.0
chromadb>=0.4.22

# Async
aiohttp>=3.9.3
asyncio>=3.4.3

# Web scraping (for website cloning)
beautifulsoup4>=4.12.3
requests>=2.31.0
selenium>=4.18.1
webdriver-manager>=4.0.1
